Extract codeblock type to separate file
