+++ src/components/SlateEditor/helpers.ts
@@ -1,6 +1,6 @@
 import { TYPE_ASIDE } from './plugins/aside/types';
 import { TYPE_BODYBOX } from './plugins/bodybox/utils';
-import { TYPE_CODEBLOCK } from './plugins/codeBlock';
+import { TYPE_CODEBLOCK } from './plugins/codeBlock/types';
 import { TYPE_CONCEPT } from './plugins/concept';
 import { TYPE_DETAILS } from './plugins/details/types';
 import { TYPE_EMBED } from './plugins/embed';
