+++ src/components/SlateEditor/plugins/codeBlock/index.tsx
@@ -15,8 +15,7 @@ import { afterOrBeforeTextBlockElement } from '../../utils/normalizationHelpers'
 import { createEmbedTag, reduceElementDataAttributes } from '../../../../util/embedTagHelpers';
 import { defaultBlockNormalizer, NormalizerConfig } from '../../utils/defaultNormalizer';
 import { TYPE_PARAGRAPH } from '../paragraph/utils';
-
-export const TYPE_CODEBLOCK = 'code-block';
+import { TYPE_CODEBLOCK } from './types';
 
 export interface CodeblockElement {
   type: 'code-block';
