+++ src/components/SlateEditor/plugins/codeBlock/types.ts
@@ -1,0 +1,1 @@
+export const TYPE_CODEBLOCK = 'code-block';
