+++ src/components/SlateEditor/plugins/codeBlock/utils.ts
@@ -1,5 +1,5 @@
 import { jsx as slatejsx } from 'slate-hyperscript';
-import { TYPE_CODEBLOCK } from '.';
+import { TYPE_CODEBLOCK } from './types';
 
 export const defaultCodeblockBlock = () =>
   slatejsx('element', { type: TYPE_CODEBLOCK, data: {}, isFirstEdit: true }, [{ text: '' }]);
